perf(explore): memoise the mini-map content-marker layers (#1015)
Every GPS fix re-renders LibreMiniMap (camera anchor, user dot and
accuracy halo genuinely track the position) — and with the content
markers inline, that re-created and re-reconciled every merchant /
cache / event / profile <Marker> for a dot that moved ~2 px: measured
145-155ms per fix on the Explore hub after #1012, one JS-thread stall
every 15s at walking pace.

The content layers move into MiniMapMarkers, a React.memo child keyed
only on the marker data + tap handlers (all already stabilised
upstream), rendered inside <MapLibreMap> — markers register via
context, so nesting is fine (CacheMapMarker precedent). uniformMarkerSize
stays a primitive prop; the derived chassis styles are computed inside
the child so the memo compare stays shallow.

After: the identical simulated walk (20 x 30m steps, Explore focused)
produces ZERO ExploreHomeScreen commits >=100ms (was 3 x 145-155ms).
Pins verified rendering + zoom controls working post-extraction.
